+++ packages/widget_toolkit/lib/src/lib_edit_address/ui_components/edit_address_widget.dart
@@ -12,21 +12,38 @@ import 'permanent_address_bottom_sheet.dart';
 /// [EditAddressLocalizedStrings], it should provided translation for the
 /// strings in the package.
 ///
-/// [editAddressErrorMapper] receives a class, which should implement
-/// [EditAddressErrorMapper] where you should map the errors
-/// for the city and street text values validation to RxFieldException<String>
-/// type, for more information check the documentation inside [EditAddressErrorMapper].
+/// [cityErrorMapper] a function, which maps the the exceptions thrown from the
+/// validation methods inside the class implementing the [TextFieldValidator]
+/// class. The methods in that class validate the input String value for the edit
+/// city text input field. The [cityErrorMapper] should map the exception to a
+/// RxFieldException
+///
+/// [addressErrorMapper] a function, which maps the the exceptions thrown from the
+/// validation methods inside the class implementing the [TextFieldValidator]
+/// class. The methods in that class validate the input String value for the edit
+/// address text input field. The [addressErrorMapper] should map the exception to a
+/// RxFieldException
+///
+/// [validator] is a service validator class, which provides methods with validation
+/// implementation for the input values of the city and address input values
+///
+/// [countryCustomIcon] receives a custom icon for the country widget. To use a
+/// custom icon, the [editCountryFieldType] should be of type
+/// EditFieldType.custom
+///
+/// [cityCustomIcon] receives a custom icon for the city button. To use a
+/// custom icon, the [editCityFieldType] should be of type
+/// EditFieldType.custom
+///
+/// [addressCustomIcon] receives a custom icon for the city button. To use a
+/// custom icon, the [editAddressFieldType] should be of type
+/// EditFieldType.custom
 ///
 /// [editAddressConfiguration] is a configuration for the edit address bottom sheet.
 ///
-/// [editAddressService] received an implementation of the [EditAddressService] class
-/// The API of the class provides methods for the logic for the main save address
-/// button, fetching of the list of countries, filtering the countries list,
-/// validating the city and street values while typing and when pressing the save
-/// button for each of them. Some methods have default implementation. For more
-/// information, check the documentation in the file [EditAddressService] class.
-///
-/// [onChanged] receives a function, which accepts the edited address model.
+/// [editAddressService] received an extension class of [EditAddressService] with
+/// implementation of the logic for the main edit contact address save button
+/// logic
 ///
 /// [editContactAddressErrorBuilder] is a custom error builder for the contact
 /// address modal sheet
@@ -37,13 +54,20 @@ import 'permanent_address_bottom_sheet.dart';
 /// picker
 class EditAddressWidget<T extends PickerItemModel> extends StatefulWidget {
   const EditAddressWidget({
-    required this.editAddressErrorMapper,
-    required this.editAddressService,
-    this.onChanged,
+    required this.translateError,
+    required this.validator,
+    required this.searchCountryService,
+    this.editAddressService = const _DefaultEditAddressService(),
     this.addressModel = _defaultAddressModel,
     this.editAddressLocalizedStrings,
     this.type = UserProfileCardTypes.mailingAddress,
     this.editAddressConfiguration = const EditAddressConfiguration(),
+    this.countryCustomIcon,
+    this.editCountryFieldType = EditFieldType.dropdown,
+    this.cityCustomIcon,
+    this.editCityFieldType = EditFieldType.editfield,
+    this.addressCustomIcon,
+    this.editAddressFieldType = EditFieldType.editfield,
     this.editContactAddressErrorBuilder,
     this.searchCountryCustomBuilders,
     this.textFieldsModalConfiguration = const TextFieldModalConfiguration(),
@@ -52,13 +76,20 @@ class EditAddressWidget<T extends PickerItemModel> extends StatefulWidget {
     Key? key,
   }) : super(key: key);
 
-  final Function(AddressModel? addressModel)? onChanged;
   final EditAddressLocalizedStrings? editAddressLocalizedStrings;
   final AddressModel addressModel;
   final UserProfileCardTypes type;
-  final EditAddressErrorMapper editAddressErrorMapper;
+  final Function(Object error) translateError;
+  final TextFieldValidator<String> validator;
   final EditAddressConfiguration editAddressConfiguration;
-  final EditAddressService<T> editAddressService;
+  final SearchPickerService<T> searchCountryService;
+  final EditAddressService editAddressService;
+  final dynamic countryCustomIcon;
+  final EditFieldType editCountryFieldType;
+  final dynamic cityCustomIcon;
+  final EditFieldType editCityFieldType;
+  final dynamic addressCustomIcon;
+  final EditFieldType editAddressFieldType;
   final Widget Function(ErrorModel?)? editContactAddressErrorBuilder;
   final SearchCountryCustomBuilders<T>? searchCountryCustomBuilders;
   final TextFieldModalConfiguration textFieldsModalConfiguration;
@@ -174,14 +205,21 @@ class _EditAddressWidgetState<T extends PickerItemModel>
       case UserProfileCardTypes.phone:
         return () async {
           final savedAddress = await showEditAddressBottomSheet<T>(context,
-              onChanged: widget.onChanged,
+              countryCustomIcon: widget.countryCustomIcon,
+              editCountryFieldType: widget.editCountryFieldType,
+              cityCustomIcon: widget.cityCustomIcon,
+              editCityFieldType: widget.editCityFieldType,
+              addressCustomIcon: widget.addressCustomIcon,
+              editAddressFieldType: widget.editAddressFieldType,
               buttonText: widget.editAddressLocalizedStrings?.saveButtonText ??
                   context.getEditAddressLocalizedStrings.saveButtonText,
               headerText: widget.editAddressLocalizedStrings?.headerTitle ??
                   context.getEditAddressLocalizedStrings.headerTitle,
               addressModel: savedModel ?? widget.addressModel,
               modalConfiguration: widget.editAddressConfiguration,
-              editAddressErrorMapper: widget.editAddressErrorMapper,
+              translateError: widget.translateError,
+              validator: widget.validator,
+              searchCountryService: widget.searchCountryService,
               editAddressLocalizedStrings: widget.editAddressLocalizedStrings,
               editAddressService: widget.editAddressService,
               editContactAddressErrorBuilder:
@@ -301,4 +339,28 @@ class SearchCountryCustomBuilders<T> {
   final Widget Function(Exception)? errorBuilder;
   final Widget Function()? emptyBuilder;
   final Widget Function(int index)? separatorBuilder;
+}
+
+class _DefaultEditAddressService implements EditAddressService {
+  const _DefaultEditAddressService();
+
+  @override
+  Future<AddressModel> saveAddress(AddressModel addressModel) async {
+    await Future.delayed(const Duration(seconds: 1));
+    throw _CustomErrorModel(
+        customMessage: 'The address could not be '
+            'saved, override the saveAddress() method and provide implementation for the '
+            'save operation.');
+  }
+}
+
+class _CustomErrorModel extends ErrorModel {
+  _CustomErrorModel({this.customMessage});
+
+  final String? customMessage;
+
+  @override
+  String toString() {
+    return customMessage ?? 'CustomErrorModel';
+  }
 }