+++ packages/widget_toolkit/test/lib_edit_address/factory/edit_address_page_factory.dart
@@ -7,7 +7,10 @@ import 'package:widget_toolkit/src/lib_edit_address/blocs/edit_address_bloc.dart
 import 'package:widget_toolkit/src/lib_edit_address/views/edit_address_page.dart';
 
 import '../mock/edit_address_mock.dart';
-import '../service/edit_address_service_mock.dart';
+import '../repository/search_country_repository_mock.dart';
+import '../service/edit_address_fields_service_mock.dart';
+import '../service/save_address_service_mock.dart';
+import '../service/search_country_service_mock.dart';
 import '../utils/edit_address_error_mapper_util.dart';
 
 /// Change the parameters according the the needs of the test
@@ -37,10 +40,12 @@ Widget editAddressPageFactory({
               address: address,
             ),
           ),
+          Provider<EditAddressFieldsServiceMock>(
+            create: (context) => EditAddressFieldsServiceMock(),
+          ),
         ],
         child: Builder(
-          builder: (context) => EditAddressPage<CountryModel>(
-            editAddressErrorMapper: EditAddressErrorMapperUtil(),
+          builder: (context) => EditAddressPage(
             addressModel: address ??
                 const AddressModel(
                   addressType: AddressTypeModel.correspondence,
@@ -49,10 +54,15 @@ Widget editAddressPageFactory({
                   country:
                       CountryModel(countryCode: 'BG', countryName: 'Bulgaria'),
                 ),
+            translateError: (obj) => EditAddressErrorMapperUtil<String>()
+                .translateError(obj, context),
+            validator: context.read<EditAddressFieldsServiceMock>(),
             buttonText: 'Save',
             headerText: 'Contact Address',
+            searchCountryService:
+                SearchCountryServiceMock(SearchCountryRepositoryMock(), false),
             editAddressService:
-                EditAddressServiceMock<CountryModel>(showError: showError ?? false),
+                SaveAddressServiceMock(showError: showError ?? false),
             editAddressLocalizedStrings: EditAddressLocalizedStrings(context),
             onAddressSaved: (AddressModel addressModel) {},
           ),
@@ -86,10 +96,12 @@ Widget createEditAddressWidget({
               address: address,
             ),
           ),
+          Provider<EditAddressFieldsServiceMock>(
+            create: (context) => EditAddressFieldsServiceMock(),
+          ),
         ],
         child: Builder(
-          builder: (context) => EditAddressWidget<CountryModel>(
-            editAddressErrorMapper: EditAddressErrorMapperUtil(),
+          builder: (context) => EditAddressWidget(
             addressModel: address ??
                 const AddressModel(
                   addressType: AddressTypeModel.correspondence,
@@ -98,7 +110,11 @@ Widget createEditAddressWidget({
                   country:
                       CountryModel(countryCode: 'BG', countryName: 'Bulgaria'),
                 ),
-            editAddressService: EditAddressServiceMock(showError:  false),
+            translateError: (obj) => EditAddressErrorMapperUtil<String>()
+                .translateError(obj, context),
+            validator: context.read<EditAddressFieldsServiceMock>(),
+            searchCountryService:
+                SearchCountryServiceMock(SearchCountryRepositoryMock(), false),
             editAddressLocalizedStrings: EditAddressLocalizedStrings(context),
           ),
         ),