+++ packages/widget_toolkit/test/lib_edit_address/utils/edit_address_error_mapper_util.dart
@@ -1,21 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_rx_bloc/rx_form.dart';
-import 'package:widget_toolkit/edit_address.dart';
 
-class EditAddressErrorMapperUtil implements EditAddressErrorMapper{
-  @override
-  RxFieldException<String> cityErrorMapper(Object error, BuildContext context) {
+class EditAddressErrorMapperUtil<T> {
+  T translateError(Object error, BuildContext context) {
     throw const RxFieldException<String>(
       error: 'Error',
-      fieldValue: 'city field value',
-    );
-  }
-
-  @override
-  RxFieldException<String> streetErrorMapper(Object error, BuildContext context) {
-    throw const RxFieldException<String>(
-      error: 'Error',
-      fieldValue: 'address error',
+      fieldValue: 'field value',
     );
   }
 }