+++ packages/widget_toolkit/test/lib_text_field_dialog/factory/text_field_dialog_factory.dart
@@ -28,8 +28,7 @@ Widget textFieldDialogFactory({
           callback: (value) => '',
           label: 'Label',
           fillButtonText: 'Save',
-          errorMapper: (Object error, BuildContext context) =>
-              const RxFieldException(
+          translateError: (Object error) => throw const RxFieldException(
             error: 'Error',
             fieldValue: 'exception',
           ),